New and hopefully final version of structs
The new strategy is to create Ensemble structs with a turbine struct array & a constants struct. Each turbine element owns its OPs and their states. Further it has its location, turbine states and the turbine type. The turbine type links to the constants where the Diameter, Efficiency and p_p for said turbine are stored. In there also the chain constants are stored